feat: solved square root

+++ 07-Binary-Search/SquareRoot.java
@@ -2,15 +2,15 @@ public class SquareRoot {
 
   public int mySqrt(int x) {
     if (x == 0 || x == 1) return x;
-    int low = 2, high = x;
+    long low = 2, high = x;
     while (low < high) {
-      int mid = low + (high - low) / 2;
+      long mid = low + (high - low) / 2;
       if (mid * mid <= x) {
         low = mid + 1;
       } else {
         high = mid;
       }
     }
-    return low - 1;
+    return (int) low - 1;
   }
 }
